feature: 커스텀 훅 작성 및 반영
